requirements.txt
pandas
pillow
tqdm
datasets
vllm
torch
litellm
python-dotenv
